+++ fluentd/vendored_gem_src/activemodel/lib/active_model/type.rb
@@ -24,15 +24,15 @@ module ActiveModel
     class << self
       attr_accessor :registry # :nodoc:
 
-      # Add a new type to the registry, allowing it to be gotten through ActiveModel::Type#lookup
-      def register(type_name, klass = nil, **options, &block)
-        registry.register(type_name, klass, **options, &block)
+      # Add a new type to the registry, allowing it to be referenced as a
+      # symbol by {attribute}[rdoc-ref:Attributes::ClassMethods#attribute].
+      def register(type_name, klass = nil, &block)
+        registry.register(type_name, klass, &block)
       end
 
-      def lookup(*args) # :nodoc:
-        registry.lookup(*args)
+      def lookup(...) # :nodoc:
+        registry.lookup(...)
       end
-      ruby2_keywords(:lookup) if respond_to?(:ruby2_keywords, true)
 
       def default_value # :nodoc:
         @default_value ||= Value.new