+++ fluentd/vendored_gem_src/activesupport/lib/active_support/core_ext/object/to_query.rb
@@ -72,14 +72,12 @@ class Hash
   #
   # The string pairs "key=value" that conform the query string
   # are sorted lexicographically in ascending order.
-  #
-  # This method is also aliased as +to_param+.
   def to_query(namespace = nil)
-    query = collect do |key, value|
+    query = filter_map do |key, value|
       unless (value.is_a?(Hash) || value.is_a?(Array)) && value.empty?
         value.to_query(namespace ? "#{namespace}[#{key}]" : key)
       end
-    end.compact
+    end
 
     query.sort! unless namespace.to_s.include?("[]")
     query.join("&")