+++ fluentd/vendored_gem_src/net-http-persistent/test/test_net_http_persistent.rb
@@ -362,6 +362,7 @@ class TestNetHttpPersistent < Minitest::Test
   end
 
   def test_connection_for_finished_ssl
+    skip 'Broken on Windows' if Gem.win_platform?
     skip 'OpenSSL is missing' unless HAVE_OPENSSL
 
     uri = URI.parse 'https://example.com/path'
@@ -555,6 +556,7 @@ class TestNetHttpPersistent < Minitest::Test
   end
 
   def test_connection_for_ssl
+    skip 'Broken on Windows' if Gem.win_platform?
     skip 'OpenSSL is missing' unless HAVE_OPENSSL
 
     uri = URI.parse 'https://example.com/path'
@@ -595,6 +597,7 @@ class TestNetHttpPersistent < Minitest::Test
   end
 
   def test_connection_for_ssl_case
+    skip 'Broken on Windows' if Gem.win_platform?
     skip 'OpenSSL is missing' unless HAVE_OPENSSL
 
     uri = URI.parse 'HTTPS://example.com/path'
@@ -631,6 +634,7 @@ class TestNetHttpPersistent < Minitest::Test
   end
 
   def test_expired_eh
+    skip 'Broken on Windows' if Gem.win_platform?
     c = basic_connection
     c.requests = 0
     c.last_use = Time.now - 11
@@ -976,7 +980,10 @@ class TestNetHttpPersistent < Minitest::Test
     assert_equal 'keep-alive', req['connection']
     assert_equal '30',         req['keep-alive']
 
-    assert_in_delta Time.now, c.last_use
+    # There's some roounding issue on jruby preventing this from passing
+    unless RUBY_PLATFORM == "java"
+      assert_in_delta Time.now, c.last_use
+    end
 
     assert_equal 1, c.requests
   end
@@ -1437,5 +1444,15 @@ class TestNetHttpPersistent < Minitest::Test
     assert_equal 1, @http.ssl_generation
   end
 
+  def test_connection_pool_after_fork
+    # ConnectionPool 2.4+ calls `checkin(force: true)` after fork
+    @http.pool.checkin(force: true)
+
+    @http.pool.checkout ['example.com', 80, nil, nil, nil, nil]
+    @http.pool.checkin(force: true)
+    @http.pool.reload do |connection|
+      connection.close
+    end
+  end
 end